functions/requirements.txt
azure-functions==1.23.0
azure-storage-blob==12.20.0
azure-storage-queue==12.10.0
pymongo==4.7.2
Pillow==10.4.0
